% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.bpnreg.R
\name{bpnreg_draw}
\alias{bpnreg_draw}
\title{Draw from Projected Normal Regression}
\usage{
bpnreg_draw(x, B)
}
\arguments{
\item{x}{A numeric matrix size N x k of covariates. (No leading 1s column.)}

\item{B}{A numeric matrix size k x 2 of regression coefficients.}
}
\value{
theta A numeric vector size N.
}
\description{
Draw from Projected Normal Regression
}
\examples{
x <- rnorm(25)
B <- matrix(c(1, 5, 0, -2), byrow = TRUE, nrow = 2)
(bpnreg_draw(x, B))
}
