% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.bpnreg.R
\name{construct_data_bpnreg}
\alias{construct_data_bpnreg}
\title{Construct the Data Frame for bpnreg}
\usage{
construct_data_bpnreg(y, ry, x)
}
\arguments{
\item{y}{A numeric vector of incomplete angles.}

\item{ry}{A logical}

\item{x}{A numeric matrix of observed covariates}
}
\value{
A data frame.
}
\description{
Construct the Data Frame for bpnreg
}
\examples{
N <- 100
X <- rnorm(N)
B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = TRUE)
y <- pnreg_draw(X, B)
mis <- sample(N, size = floor(0.5 * N), replace = FALSE)
y[mis] <- NA
ry <- !is.na(y)
x <- cbind(rep(1, N), X)
(construct_data_bpnreg(y, ry, x))
}
