% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.bpnreg.R
\name{convert2MinusPiPlusPi}
\alias{convert2MinusPiPlusPi}
\title{Convert Angle to Minus Pi Plus Pi Interval}
\usage{
convert2MinusPiPlusPi(theta)
}
\arguments{
\item{theta}{A numeric vector of angular values.}
}
\value{
A numeric vector.
}
\description{
Convert Angle to Minus Pi Plus Pi Interval
}
\examples{
theta <- seq(0, 2*pi, length.out = 12)
(convert2MinusPiPlusPi(theta))
}
