% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.bpnreg.R
\name{get_posterior_draws_pnreg}
\alias{get_posterior_draws_pnreg}
\title{Collect Posterior Draws for Coefficients}
\usage{
get_posterior_draws_pnreg(fit)
}
\arguments{
\item{fit}{A fitted object from bpnreg package.}
}
\value{
A numeric matrix size k x 2 of regression coefficients
}
\description{
Collect Posterior Draws for Coefficients
}
\examples{
N <- 100
x <- rnorm(N)
B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = TRUE)
y <- pnreg_draw(x, B)
ry <- !is.na(y)
dat <- construct_data_bpnreg(y, ry, x)
invisible(
    utils::capture.output(
        fit <- bpnreg::bpnr(theta ~ ., data = dat, its = 2000, burn = 1000)
        )
    )
(get_posterior_draws_pnreg(fit))
}
