% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_pgp_geostat_reg.R
\name{impute_pgpreginc}
\alias{impute_pgpreginc}
\title{Impute with Projected Gaussian Process Geostatical Model with Internally Imputed Angles}
\usage{
impute_pgpreginc(loc, theta, x, M, ...)
}
\arguments{
\item{loc}{matrix of location coordinates (N x 2)}

\item{theta}{angles}

\item{x}{predictors}

\item{M}{number of imputations}

\item{...}{other arguments to pass to fit_pgp_geostat_reg_inc_model()}
}
\value{
A numeric vector of length n_mis.
}
\description{
The projected Gaussian process geostatistical model with missing angular
observations imputed during the sampling procedure. The imputation step
integrates out the missingness from the model fitting. Imputes only on
the incomplete angular data.
}
\examples{
## Not run:
  loc <- expand.grid(long = 1:10, lat = 1:10)[,1:2]
  loc2 <- expand.grid(long = 5, lat = 5)[,1:2]
  X <- pnregstan::gp_geostat_sim_data(N = 10, M = 10, sigma = 1, rho = 5, alpha = 1,
                        refresh = 0,
                        show_messages = FALSE,
                        show_exceptions = FALSE)$X
  X <- as.matrix(X, nrow = 100)

  df <- pnregstan::pgp_geostat_reg_sim_data(loc = loc, X = X, sigma_w = 1, rho_w = 0, rho = 5,
         mu0 = c(1,0), B = matrix(c(-0.5, 1), nrow = 1),
         refresh = 0,
         show_messages = FALSE,
         show_exceptions = FALSE)
  ind_miss <- 1:25 * 4

  theta <- df$theta
  theta[ind_miss] <- NA
  rt <- !is.na(theta)

  impute_pgpreginc(loc = loc, theta = theta, x = X, M = 2, iter_warmup = 10, iter_sampling = 10)

## End(Not run)
}
