% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.bpnreg.R
\name{mice.impute.bpnreg}
\alias{mice.impute.bpnreg}
\title{Impute with Projected Normal Regression}
\usage{
mice.impute.bpnreg(y, ry, x, ...)
}
\arguments{
\item{y}{The incomplete vector of N angles to be imputed.}

\item{ry}{A logical vector of length N indicating whether the ith observation is observed (TRUE) or missing (FALSE).}

\item{x}{Completely observed covariates with dimension N x k.}

\item{...}{Additional parameters (unused)}
}
\value{
A numeric vector of length n_mis.
}
\description{
Impute with Projected Normal Regression
}
\examples{
N <- 100
X <- rnorm(N)
theta <- numeric(N)
for (i in 1:N) {
    B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = T)
    mu_i <- c(1, X[i]) \%*\% B
    Y <- mvtnorm::rmvnorm(N, mu_i, diag(2))
    U <- Y / sqrt(sum(Y^2))
    theta[i] <- atan2(U[2], U[1])
}
mis <- sample(N, size = floor(0.5 * N), replace = FALSE)
theta[mis] <- NA
ry <- !is.na(theta)
X1 <- cbind(rep(1, N), X)

mice.impute.bpnreg(theta, ry, X1)
}
