% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pnarx.R
\name{mice.impute.pnarx}
\alias{mice.impute.pnarx}
\title{Impute with Projected Normal AR with Exogenous Predictors}
\usage{
mice.impute.pnarx(y, ry, x, ...)
}
\arguments{
\item{y}{angles}

\item{ry}{response indicators}

\item{x}{predictors}

\item{...}{other arguments to pass to fit_pn_arx_model()}
}
\value{
A numeric vector of length n_mis.
}
\description{
Impute with Projected Normal AR with Exogenous Predictors
}
\examples{
## Not run:
N <- 100
x <- rnorm(N)
B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = TRUE)
y <- bpnreg_draw(x, B)
mis <- sample(N, size = floor(0.5 * N), replace = FALSE)
y[mis] <- NA
ry <- !is.na(y)

mice.impute.pnarx(y, ry, x)
## End(Not run)
}
