% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pnarxinc.R
\name{mice.impute.pnarxinc}
\alias{mice.impute.pnarxinc}
\title{Impute with Projected Normal AR with Exogenous Predictors (Identity Covariance) and Internally Imputed Angles}
\usage{
mice.impute.pnarxinc(y, ry, x, ...)
}
\arguments{
\item{y}{angles}

\item{ry}{response indicators}

\item{x}{predictors}

\item{...}{other arguments to pass to fit_pn_arx_model()}
}
\value{
A numeric vector of length n_mis.
}
\description{
The projected normal AR with exogenous predictors and identity covariance is
with missing angular observations imputed during the sampling procedure. This
integrates out the missingness from the model fitting so if missingness is only
on theta, then only one cycle should be necessary.
}
\examples{
## Not run:
\dontrun{
N <- 100
x <- rnorm(N)
B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = TRUE)
y <- bpnreg_draw(x, B)
mis <- sample(N, size = floor(0.25 * N), replace = FALSE)
y[mis] <- NA
ry <- !is.na(y)

mice.impute.pnarxid(y, ry, x)
}
## End(Not run)
}
