% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.vmreg.R
\name{mice.impute.vmreg}
\alias{mice.impute.vmreg}
\title{Impute Incomplete Angular Data with von Mises Regression}
\usage{
mice.impute.vmreg(y, ry, x, ...)
}
\arguments{
\item{y}{The incomplete vector of N angles to be imputed.}

\item{ry}{A logical vector of length N indicating whether the ith observation is observed (TRUE) or missing (FALSE).}

\item{x}{Completely observed covariates with dimension N x k.}

\item{...}{Additional parameters (unused)}
}
\value{
A numeric vector of length n_mis.
}
\description{
Impute Incomplete Angular Data with von Mises Regression
}
\examples{
N <- 100
x <- rnorm(N)
B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = TRUE)
y <- pnreg_draw(x, B)
mis <- sample(N, size = floor(0.5 * N), replace = FALSE)
y[mis] <- NA
ry <- !is.na(y)

mice.impute.vmreg(y, ry, x)
}
