% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_angular_imputations.R
\name{plot_angular_imputations}
\alias{plot_angular_imputations}
\title{Plot Angular Imputations}
\usage{
plot_angular_imputations(imps, r = 0.25, alpha = 0.5, by_id = FALSE)
}
\arguments{
\item{imps}{A mids object produced by mice}

\item{r}{numeric Radius for unit circles}

\item{alpha}{numeric opacity for the points}

\item{by_id}{bool Switch for plots of unit circles for each imputation or plots of vectors by id and by imputation}
}
\value{
A ggplot object
}
\description{
Create a plot of the angular imputations produced by mice. The plot displays a unit circle for each imputation with observed points plotted as blue points on the circle and imputations plotted as pink points on the circle. This plot is similar in nature to the stripplot included in the mice package.
}
\examples{
   N <- 30
}
