% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.vmreg.R
\name{ppd_draws_vmreg}
\alias{ppd_draws_vmreg}
\title{Collect Posterior Predictive Draws for von Mises Regression}
\usage{
ppd_draws_vmreg(fit, ry, x)
}
\arguments{
\item{fit}{A 'circglm' fit object.}

\item{ry}{A logical vector of length N indicating whether the ith observation is observed (TRUE) or missing (FALSE).}

\item{x}{A N x k centered and scaled predictor matrix.}
}
\value{
theta_imp A numeric vector of size n_mis.
}
\description{
Collect Posterior Predictive Draws for von Mises Regression
}
\examples{
N <- 100
x <- rnorm(N)
B <- matrix(c(3.5, 1, -3, 0), nrow = 2, byrow = TRUE)
y <- bpnreg_draw(x, B)
mis <- sample(N, size = floor(0.5 * N), replace = FALSE)
y[mis] <- NA
ry <- !is.na(y)
xc <- scale(x)
dat <- construct_modeling_data(y, ry, xc)
fit <- circglmbayes::circGLM(theta ~ ., data = dat,
                             Q = 100, thin = 10, burnin = 25000,
                             bt_prior_musd = c(mu = 0, sd = 1),
                             r = 2)
ppd_draws_vmreg(fit, ry, x)
}
